move client/msg_service.go to types/module/msg_service.go and move util/utils.go to x/ecocredit/util/utils.go
