+++ types/go.mod
@@ -16,5 +16,3 @@ require (
 )
 
 replace github.com/gogo/protobuf => github.com/regen-network/protobuf v1.3.3-alpha.regen.1
-
-replace github.com/regen-network/regen-ledger => ../
