+++ x/group/go.mod
@@ -9,16 +9,14 @@ require (
 	github.com/gorilla/mux v1.8.0
 	github.com/grpc-ecosystem/grpc-gateway v1.16.0
 	github.com/regen-network/cosmos-proto v0.3.1
-	github.com/regen-network/regen-ledger v1.0.0
 	github.com/regen-network/regen-ledger/orm v0.0.0-00010101000000-000000000000
 	github.com/regen-network/regen-ledger/types v0.0.0-00010101000000-000000000000
 	github.com/spf13/cobra v1.1.3
 	github.com/stretchr/testify v1.7.0
 	github.com/tendermint/tendermint v0.34.10
+	github.com/tendermint/tm-db v0.6.4
 	google.golang.org/grpc v1.37.0
 )
-
-replace github.com/regen-network/regen-ledger => ../../
 
 replace github.com/regen-network/regen-ledger/orm => ../../orm
